Lesson 2 Updated notes
Lesson 2 Updated notes:

Added in notes in the total-spent-by-customer.py to show how the reducebykey() function really works by:

In this specific example the reducebykey function takes a lambda function. It takes 2 arguments, which represents two values that have the SAME key so we calculate the sum of the two values and return it to a single value to the SAME key.

and how to sort a key, value pair by only sorting through the key by flipping the pair.