Lesson 4
Lesson 4:

Finished lesson 4 about using BFS breadth-first-search to find the distance from one connection to another or degrees of separation. Also learned about sparks cache and persist functions. NOTE: that I was unable to completely learn and run the movie similarities due to some issue and need to come back to it later.